Update Answer 6

+++ 0x00-hello_world/6-size.c
@@ -11,14 +11,14 @@ int main(void)
 {
 	char c;
 	int b;
-	unsigned int b1;
-	long d;
+	long int b1;
+	long long int d;
 	float d1;
 
-	printf("Size of a char : %li byte(s)\n", sizeof(c));
-	printf("Size of a int : %li byte(s)\n", sizeof(b));
-	printf("Size of a long int : %li byte(s)\n", sizeof(b1));
-	printf("Size of a long long int: %li byte(s)\n", sizeof(d));
-	printf("Size of a float : %li byte(s)\n", sizeof(d1));
+	printf("Size of a char : %i byte(s)\n", sizeof(c));
+	printf("Size of a int : %i byte(s)\n", sizeof(b));
+	printf("Size of a long int : %i byte(s)\n", sizeof(b1));
+	printf("Size of a long long int: %i byte(s)\n", sizeof(d));
+	printf("Size of a float : %i byte(s)\n", sizeof(d1));
 	return (0);
 }
